+++ pubspec.yaml
@@ -31,12 +31,16 @@ dependencies:
   flutter:
     sdk: flutter
 
-
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   # The following is used for @protected
   meta: 1.8.0
+  # BLoC state management
+  bloc: ^8.0.3
+  bloc_test: ^9.0.3
+  # Library that helps with class equality
+  equatable: ^2.0.3
 
 dev_dependencies:
   test: ^1.21.4
