+++ test/widget_test.dart
@@ -7,14 +7,12 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:incremental_increment/game/engine/util/number.dart';
-
-import 'package:incremental_increment/main.dart';
+import 'package:incremental_increment/antimatter_dimensions/components/antimatter_dimensions_widget.dart';
 
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
     // Build our app and trigger a frame.
-    await tester.pumpWidget(const MyApp());
+    await tester.pumpWidget(const AntimatterDimensionsApp());
 
     // Verify that our counter starts at 0.
     expect(find.text('0'), findsOneWidget);
